backend/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
applescript==2021.2.9
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
audioop-lts==0.2.1
babel==2.17.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.1.8
cloudpickle==3.1.1
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.2
cryptography==45.0.3
curl_cffi==0.11.1
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
distro==1.9.0
durationpy==0.10
elevenlabs==2.4.0
environment==1.0.0
executing==2.2.0
Farama-Notifications==0.0.4
fastapi==0.115.12
fastjsonschema==2.21.1
ffmpy==0.6.0
filelock==3.18.0
filetype==1.2.0
Flask==3.1.1
flask-cors==6.0.1
flatbuffers==25.2.10
fonttools==4.58.1
fqdn==1.5.1
frozendict==2.4.6
frozenlist==1.6.0
fsspec==2025.5.0
google==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.170.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-genai==1.7.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
gradio==5.34.0
gradio_client==1.10.3
groovy==0.1.2
grpcio==1.72.0rc1
grpcio-status==1.71.0
gym==0.26.2
gym-notices==0.0.8
gymnasium==1.1.1
h11==0.14.0
hf-xet==1.1.4
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.31.4
humanfriendly==10.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
json5==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
kubernetes==32.0.1
langchain==0.3.26
langchain-community==0.3.24
langchain-core==0.3.66
langchain-google-genai==2.1.2
langchain-tavily==0.1.6
langchain-text-splitters==0.3.8
langgraph==0.3.21
langgraph-checkpoint==2.0.26
langgraph-prebuilt==0.1.7
langgraph-sdk==0.1.69
langsmith==0.4.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
mmh3==5.1.0
mpmath==1.3.0
multidict==6.4.4
multitasking==0.0.11
mypy==1.15.0
mypy_extensions==1.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nbimporter==0.3.4
nest-asyncio==1.6.0
networkx==3.5
notebook_shim==0.2.4
numpy==2.2.3
oauthlib==3.2.2
onnxruntime==1.22.0
opentelemetry-api==1.33.1
opentelemetry-exporter-otlp-proto-common==1.33.1
opentelemetry-exporter-otlp-proto-grpc==1.33.1
opentelemetry-instrumentation==0.54b1
opentelemetry-instrumentation-asgi==0.54b1
opentelemetry-instrumentation-fastapi==0.54b1
opentelemetry-proto==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
opentelemetry-util-http==0.54b1
orjson==3.10.18
ormsgpack==1.9.1
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pdfminer.six==20250506
peewee==3.18.1
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.7
posthog==4.0.1
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.1
pydantic-settings==2.9.1
pydantic_core==2.33.0
pydub==0.25.1
Pygments==2.19.1
PyMuPDF==1.26.1
pyparsing==3.2.3
pypdf==5.6.0
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.3.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.25.0
rsa==4.9
ruff==0.12.0
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
semantic-version==2.10.0
Send2Trash==1.8.3
sentence-transformers==4.1.0
setuptools==80.7.1
shellingham==1.5.4
Shimmy==2.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
spark-parser==1.8.9
SQLAlchemy==2.0.41
stable_baselines3==2.6.0
stack-data==0.6.3
starlette==0.46.2
sympy==1.14.0
tenacity==9.1.2
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tokenizers==0.21.1
tomlkit==0.13.3
torch==2.7.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.52.4
typer==0.15.4
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.1
uncompyle6==3.9.2
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
widgetsnbextension==4.0.14
wrapt==1.17.2
xdis==6.1.1
xxhash==3.5.0
yarl==1.20.0
yfinance==0.2.61
zipp==3.21.0
zstandard==0.23.0
